@@ -2,7 +2,7 @@ name: giant_gipsland_earthworm_fe
 description: "A new Flutter project."
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+68
+version: 1.0.2+91
 
 environment:
   sdk: ">=2.19.1 <3.7.3"
@@ -13,62 +13,62 @@ dependencies:
 
   cupertino_icons: ^1.0.6
   get: ^4.6.6
-  firebase_auth: ^5.3.1
-  dio: ^5.5.0
+  firebase_auth: ^5.4.2
   connectivity_plus: ^4.0.1
   go_router: ^14.2.0
   flutter_svg: ^2.0.10+1
   fluttertoast: ^8.2.6
   toastification: ^2.0.0
   loading_animation_widget: ^1.2.1
-  shimmer: ^3.0.0
   package_info_plus: ^8.0.0
   form_field_validator: ^1.1.0
-  cloud_firestore:
+  cloud_firestore: ^5.6.3
   google_sign_in: ^6.2.1
   image_picker: ^1.1.2
-  firebase_storage: ^12.3.5
+  firebase_storage: ^12.4.2
   cached_network_image: ^3.3.1
   calendar_date_picker2: ^1.1.5
   geolocator: ^12.0.0
   geocoding: ^3.0.0
   flutter_markdown: ^0.7.3
   uuid: ^4.4.0
-  firebase_analytics: ^11.3.3
-  flutter_facebook_auth: ^7.0.1
-  flutter_launcher_icons: ^0.14.1
-  google_maps_flutter: ^2.6.1
+  firebase_analytics: ^11.4.2
+  flutter_launcher_icons: ^0.13.0
+  google_maps_flutter: ^2.10.0
   photo_view: ^0.15.0
   http:
   saver_gallery: ^3.0.5
   permission_handler: ^11.3.1
-  flutter_typeahead: ^5.2.0
   the_apple_sign_in: ^1.1.1
   url_launcher: ^6.3.0
   record: ^5.1.2
   audioplayers: ^6.0.0
   readmore: ^3.0.0
-  file_picker: ^8.0.5
-  csv: ^6.0.0
-  path: ^1.9.0
-  device_preview: ^1.1.0
+  file_picker: ^9.0.0
+  # path: ^1.9.0
   video_player: ^2.9.2
-  flick_video_player: ^0.9.0
-  animated_loading_border: ^0.0.2
-  firebase_core: ^3.7.0
+  firebase_core: ^3.11.0
   path_provider: ^2.1.5
-  # audioplayers: ^6.0.0
-  # flutter_sound_record: ^3.3.2
+  appinio_video_player: ^1.2.2
+  get_storage: ^2.1.1
+  lottie: 
+  flutter_cache_manager: ^3.4.1
+  hive: ^2.2.3
+  firebase_crashlytics: ^4.3.5
+  flick_video_player: ^0.9.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   flutter_lints: ^3.0.0
+  hive_generator: ^2.0.1
+  build_runner: ^2.4.6
   # build_runner: ^2.4.11
 
 dependency_overrides:
-  win32: ^5.5.4
+  win32: 
+  image: ^3.0.2
 
 flutter:
   uses-material-design: true
@@ -78,6 +78,7 @@ flutter:
     - assets/icons/
     - assets/logo/
     - assets/commonassets/
+    - assets/json/
 
   fonts:
     - family: BeVietnamPro
