@@ -1,3 +1,5 @@
+import 'package:giant_gipsland_earthworm_fe/features/dashboard/home_screen/model/survey_model.dart';
+
 class SampleVideoModel {
   String img;
   String videoLink;
@@ -50,29 +52,41 @@ class SampleVideoModel {
 class HomeScreenDataModel {
   List<SampleVideoModel> sampleVideos;
   String howDataWillBeUsed;
+  List<SurveyModel> surveys; // New field for surveys
 
   HomeScreenDataModel({
     required this.sampleVideos,
     required this.howDataWillBeUsed,
+    required this.surveys,
   });
 
-  // To Map (for Firestore)
+  // Convert to Map for Firestore
   Map<String, dynamic> toMap() {
     return {
       'sampleVideos': sampleVideos.map((video) => video.toMap()).toList(),
       'how_data_will_be_used': howDataWillBeUsed,
+      'surveys': surveys
+          .map((survey) => survey.toJson())
+          .toList(), // Convert surveys to JSON
     };
   }
 
-  // From Firestore
+  // Create an instance from Firestore data
   factory HomeScreenDataModel.fromMap(Map<String, dynamic> map) {
     return HomeScreenDataModel(
       sampleVideos: List<SampleVideoModel>.from(
-        (map['sampleVideos'] as List<dynamic>?)?.map((item) =>
-                SampleVideoModel.fromMap(item as Map<String, dynamic>)) ??
+        (map['sampleVideos'] as List<dynamic>?)?.map(
+              (item) => SampleVideoModel.fromMap(item as Map<String, dynamic>),
+            ) ??
             [],
       ),
       howDataWillBeUsed: map['how_data_will_be_used'] ?? '',
+      surveys: List<SurveyModel>.from(
+        (map['surveys'] as List<dynamic>?)?.map(
+              (item) => SurveyModel.fromJson(item as Map<String, dynamic>),
+            ) ??
+            [],
+      ),
     );
   }
 
@@ -81,6 +95,7 @@ class HomeScreenDataModel {
     return HomeScreenDataModel(
       sampleVideos: [],
       howDataWillBeUsed: '',
+      surveys: [],
     );
   }
 }
