@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart'; // GetX for state management and navigation
+import 'package:giant_gipsland_earthworm_fe/core/common_controller/internet_check_controller.dart';
 import 'package:giant_gipsland_earthworm_fe/core/common_widgets/alert_msg_dialog.dart'; // Custom dialog widget
+import 'package:giant_gipsland_earthworm_fe/core/common_widgets/common_toast.dart';
 import 'package:giant_gipsland_earthworm_fe/core/constants/app_constants.dart'; // Constants like padding
 import 'package:giant_gipsland_earthworm_fe/core/constants/app_image_constant.dart'; // Constants for image paths
 import 'package:giant_gipsland_earthworm_fe/core/custom_route_utills/custom_navigation.dart'; // Custom navigation helper
@@ -16,8 +18,8 @@ class AccountSettingScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Padding(
-        padding: AppConstants
-            .screenPadding(), // Custom screen padding from constants
+        padding: AppConstants.screenPadding(
+            context: context), // Custom screen padding from constants
         child: SingleChildScrollView(
           // Allows scrolling if content overflows
           child: Column(
@@ -41,20 +43,31 @@ class AccountSettingScreen extends StatelessWidget {
 
               // Menu item for deleting account
               ProfileMenuContainer(
-                text: "Delete Account",
-                ontap: () => Get.dialog(CommonAlertMessageDialog(
-                    description:
-                        "Are you sure you want to delete your account? This action cannot be undone, and all your data will be permanently removed.",
-                    title: "Delete Account?",
-                    icon: AppImagesConstant
-                        .deleteUserIconPNG, // Icon for deletion
-                    action: () {
-                      Get.back(); // Close the dialog
-                      ChangePasswordController
-                          .instance // Call method to delete account
-                          .deleteAccount(context: context);
-                    })),
-              ),
+                  text: "Delete Account",
+                  ontap: () {
+                    if (InternetCheckController.instance.isConnected.value ==
+                        false) {
+                      showCommonToast(
+                        context: context,
+                        title: "Internet Connection Required",
+                        description:
+                            "Please connect to the internet and try again.",
+                      );
+                    } else {
+                      Get.dialog(CommonAlertMessageDialog(
+                          description:
+                              "Are you sure you want to delete your account? This action cannot be undone, and all your data will be permanently removed.",
+                          title: "Delete Account?",
+                          icon: AppImagesConstant
+                              .deleteUserIconPNG, // Icon for deletion
+                          action: () {
+                            Get.back(); // Close the dialog
+                            ChangePasswordController
+                                .instance // Call method to delete account
+                                .deleteAccount(context: context);
+                          }));
+                    }
+                  }),
             ],
           ),
         ),
