@@ -3,12 +3,12 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:giant_gipsland_earthworm_fe/core/constants/app_constants.dart';
-import 'package:giant_gipsland_earthworm_fe/core/constants/app_image_constant.dart';
 import 'package:giant_gipsland_earthworm_fe/core/constants/common_assets.dart';
 import 'package:giant_gipsland_earthworm_fe/features/dashboard/add_worms/model/worm_model.dart';
 import 'package:giant_gipsland_earthworm_fe/features/dashboard/add_worms/widgets/playing_audio_widget.dart';
@@ -48,6 +48,7 @@ class WormDetailsScreen extends StatelessWidget {
     return Scaffold(
       body: Padding(
         padding: AppConstants.screenPadding(
+            context: context,
             androidBottom: 0), // Screen padding for layout consistency
         child: Column(
           children: [
@@ -118,86 +119,85 @@ class WormDetailsScreen extends StatelessWidget {
                         style: textTheme.titleMedium?.copyWith(fontSize: 20)),
                     const SizedBox(height: 5),
                     // FutureBuilder to fetch and display audio if available
-                    FutureBuilder(
-                      future: CommonAssets.getGCSUrl(wormDetails
-                          .audioUrl), // Fetching the audio URL from GCS
-                      builder: (context, snapshot) {
-                        if (snapshot.connectionState ==
-                            ConnectionState.waiting) {
-                          return const Center(
-                              child:
-                                  CircularProgressIndicator()); // Loading state
-                        } else if (snapshot.hasError) {
-                          return Center(
-                              child: Text(
-                                  'Error: ${snapshot.error}')); // Error state
-                        } else if (snapshot.hasData) {
-                          return AudioPlayerWidget(
-                              audioUrl: snapshot
-                                  .data!); // Display audio player widget
-                        } else {
-                          return Center(
-                            child: Text('No audio available',
-                                style: textTheme.titleMedium
-                                    ?.copyWith(fontSize: 18)), // No audio state
-                          );
-                        }
-                      },
-                    ),
+                    wormDetails.audioUrl.isEmpty
+                        ? Text('No audio available',
+                            style: textTheme.titleMedium
+                                ?.copyWith(fontSize: 18, color: Colors.red))
+                        : FutureBuilder(
+                            future: CommonAssets.getGCSUrl(wormDetails
+                                .audioUrl), // Fetching the audio URL from GCS
+                            builder: (context, snapshot) {
+                              if (snapshot.connectionState ==
+                                  ConnectionState.waiting) {
+                                return const Center(
+                                    child:
+                                        CircularProgressIndicator()); // Loading state
+                              } else if (snapshot.hasError) {
+                                return Center(
+                                    child: Text(
+                                        'Error: ${snapshot.error}')); // Error state
+                              } else if (snapshot.hasData) {
+                                return AudioPlayerWidget(
+                                    audioUrl: snapshot
+                                        .data!); // Display audio player widget
+                              } else {
+                                return Text('No audio available',
+                                    style: textTheme.titleMedium?.copyWith(
+                                        fontSize: 18, color: Colors.red));
+                              }
+                            },
+                          ),
                     const SizedBox(height: 5),
                     // Location image section header
                     Text("Location Image",
                         style: textTheme.titleMedium?.copyWith(fontSize: 20)),
                     const SizedBox(height: 5),
                     // Display location image in a clickable widget
-                    ClipRRect(
-                      borderRadius: BorderRadius.circular(8),
-                      child: InkWell(
-                        onTap: () => Get.to(() => FullScreenImage(
-                            imageUrl: wormDetails
-                                .locationImg)), // Open full-screen image view
-                        child: Hero(
-                          tag: wormDetails.locationImg, // Hero animation tag
-                          child: CommonAssets.getGCSNetworkImage(
-                              wormDetails.locationImg,
-                              AppImagesConstant.appLogo,
-                              height: 130,
-                              width: double.infinity,
-                              fit: BoxFit
-                                  .cover), // Display image fetched from GCS
-                        ),
-                      ),
-                    ),
+                    wormDetails.locationImg.isEmpty
+                        ? Text('No Location Image available',
+                            style: textTheme.titleMedium
+                                ?.copyWith(fontSize: 18, color: Colors.red))
+                        : ClipRRect(
+                            borderRadius: BorderRadius.circular(8),
+                            child: InkWell(
+                              onTap: () => Get.to(() => FullScreenImage(
+                                  imageUrl: wormDetails
+                                      .locationImg)), // Open full-screen image view
+                              child: Hero(
+                                tag: wormDetails
+                                    .locationImg, // Hero animation tag
+                                child: GCSCachedImage(
+                                  gcsPath: wormDetails.locationImg,
+                                  height: 130,
+                                  width: double.infinity,
+                                  // fit: BoxFit
+                                  //     .cover
+                                ), // Display image fetched from GCS
+                              ),
+                            ),
+                          ),
                     const SizedBox(height: 10),
                     // Worm images section header
                     Text("Worm Images",
                         style: textTheme.titleMedium?.copyWith(fontSize: 20)),
                     const SizedBox(height: 5),
                     // Displaying multiple worm images in a wrap widget
-                    Wrap(
-                      runSpacing: 10,
-                      spacing: 10,
-                      children: wormDetails.wormsImg
-                          .map((e) => ClipRRect(
-                                borderRadius: BorderRadius.circular(8),
-                                child: InkWell(
-                                  onTap: () => Get.to(() => FullScreenImage(
-                                      imageUrl:
-                                          e)), // Open full-screen image view
-                                  child: Hero(
-                                    tag: e
-                                        .characters, // Hero animation tag for each image
-                                    child: CommonAssets.getGCSNetworkImage(
-                                        e, AppImagesConstant.appLogo,
-                                        height: 100,
-                                        width: 100,
-                                        fit: BoxFit
-                                            .cover), // Display worm images
-                                  ),
-                                ),
-                              ))
-                          .toList(),
-                    ),
+                    wormDetails.wormsImg.isEmpty
+                        ? Text('No Worms Image available',
+                            style: textTheme.titleMedium
+                                ?.copyWith(fontSize: 18, color: Colors.red))
+                        : SizedBox(
+                            height: 120, // Adjust as needed
+                            child: ListView.builder(
+                              scrollDirection: Axis.horizontal,
+                              itemCount: wormDetails.wormsImg.length,
+                              itemBuilder: (context, index) {
+                                final gcsUrl = wormDetails.wormsImg[index];
+                                return GCSCachedImage(
+                                    gcsPath: gcsUrl, height: 100, width: 100);
+                              },
+                            ),
+                          ),
                     const SizedBox(
                         height: 20), // Space at the bottom of the page
                   ],
@@ -207,6 +207,73 @@ class WormDetailsScreen extends StatelessWidget {
           ],
         ),
       ),
+    );
+  }
+}
+
+class GCSCachedImage extends StatelessWidget {
+  final String gcsPath;
+  final double height;
+  final double width;
+  final double borderRadius;
+  final EdgeInsetsGeometry padding;
+
+  const GCSCachedImage({
+    super.key,
+    required this.gcsPath,
+    this.height = 100,
+    this.width = 100,
+    this.borderRadius = 8,
+    this.padding = const EdgeInsets.only(right: 10),
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    if (gcsPath.isEmpty) {
+      return const SizedBox(
+        width: 100,
+        height: 100,
+        child: Center(child: Icon(Icons.image_not_supported)),
+      );
+    }
+
+    return FutureBuilder<String?>(
+      future: CommonAssets.getGCSUrl(gcsPath),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return SizedBox(
+            width: width,
+            height: height,
+            child: const Center(child: CircularProgressIndicator()),
+          );
+        } else if (snapshot.hasError || !snapshot.hasData) {
+          return SizedBox(
+            width: width,
+            height: height,
+            child: const Icon(Icons.error),
+          );
+        }
+
+        final publicUrl = snapshot.data!;
+        return Padding(
+          padding: padding,
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(borderRadius),
+            child: InkWell(
+              onTap: () => Get.to(() => FullScreenImage(imageUrl: publicUrl)),
+              child: CachedNetworkImage(
+                imageUrl: publicUrl,
+                width: width,
+                height: height,
+                fit: BoxFit.cover,
+                placeholder: (context, url) =>
+                    const Center(child: CircularProgressIndicator()),
+                errorWidget: (context, url, error) => const Icon(Icons.error),
+              ),
+            ),
+          ),
+        );
+      },
     );
   }
 }
